+++ src/tools/chaincodeinventory.ts
@@ -15,10 +15,10 @@ export namespace ChaincodeInventory {
 
     try {
       logger.debug('Creating chaincode inventory...');
-      const _inv = yaml.safeLoad(fs.readFileSync(Config.getConfigItem("inventory_file"), 'utf8'));
-      _inventory = _inv.ccMethods;
+      const _inv = yaml.load(fs.readFileSync(Config.getConfigItem("inventory_file"), 'utf8'));
+      _inventory = _inv["ccMethods"];
   
-      Object.values(_inv.ccMethods).forEach((_item: any) => {
+      Object.values(_inv["ccMethods"]).forEach((_item: any) => {
         const _invItem: FabricConfig = _item;
         // This transformation is needed because of YAML label substitution
         _invItem.channelName = _item.channelName.name;