+++ CHANGELOG.md
@@ -7,14 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added 
+## [0.2.0] - 2021-0?-??
 
-- Add `enrollUser` utilitiy
+### Added
+
+- Add `createNewUser` utilitiy
+- Add `getConnectedGateway` utility
+- Add default users initialization logic
 
 ### Changed
 
 - Replace the manual type definitions with `tsc` generated one
 - Replace `mocha` testing framework with `jest`
+- Update the `fabric` dependencies from `v1.4` to `v2.2`
+- Update the fabric user registration and enrollment logic to `v2.2`
 
 ## [0.1.4] - 2020-09-22